extender: add an iterator adapter for generating Buckets
Our overall plan is for `SortBuf` to be constructed from a heap of
buckets of (more or less) identical size. Obviously, an arbitrary source
of items will not give us pre-packaged buckets. It's our job to produce
these buckets.

This iterator adapter allows us to do so, without depending on the
source of items to come from a single iterator. The latter will prove to
be rather important when we implement `Extend` for user-facing types in
the future.