bucket: add type for transporting buckets through user-facing API
Instead of introducing functions to `SortBuf` itself for adding items,
we decided for a more indirect appraoch. It will involve an "extender"
based on a trait for adding buckets. We plan to implement that trait not
only for `SortBuf` (or mutable references), but also other related types
in order to cover more (interesting) use-cases.

This trait will ultimately require making buckets visible to users in
some way. However, we don't want to expose too many internas such as the
implementation of `Ord` for `SortedBucket`. The type introduced by this
change serves the purpose of transporting bucket data in a way that is
sufficiently opaque to the user.