extender: add trait for bucket accumulators
Instead of introducing functions to `SortBuf` itself for adding items,
we decided for a more indirect appraoch. It will involve an "extender"
based on a trait for adding buckets. We plan to implement that trait not
only for `SortBuf` (or mutable references), but also other related types
in order to cover more (interesting) use-cases.

This change introduces this trait. The observant reader will notice that
its definition is awfully close to `Extend`. We normally would prefer to
re-use an existing trait. However, we cannot implement `Extend` for all
types we would like to. In particular, we cannot implement it for
smart-ponter-like types wrapping a `SortBuf`.