extender: add a type for extending a BucketAllocator by items
Our overall plan is for `SortBuf` to be constructed from a heap of
buckets of (more or less) identical size. Obviously, an arbitrary source
of items will not give us pre-packaged buckets. It's our job to produce
these buckets.

For this purpose, we recently introduced the `BucketAccumulator` trait.
This change introduces the last missing piece, a type which lets users
add items to implementers of that trait, including `SortBuf`.