extender: add a fn for creating an Extender with a default bucket size
We don't want to force our users to think about a reasonable target size
for buckets. This convenience function offers a reasonable [TM] default.

Users who may want to set another bucket size can do so via previously
introduced functions. However, we suspect that at least some users will
want to play around with bucket sizes relative to the default size. We
therefore decided to expose this value, albeit we may make it private in
the future if we feel like it.