extender: make BucketAccumulator accept individual Buckets
... rather than an iterator over `Bucket`s.

The previous approach didn't capture the intend to keep computation on
the callers side, which is important especially in multithreading
use-cases, e.g. when putting a `SortBuf` behind a mutex. In this case,
we'd keep the mutex locked while iterating over the item source, which
could be arbitrarily expensive. With this change, we keep the lock only
while pushing a bucket, not while accumulating items into one.