bucket: make Bucket construction responsible for sorting items
Previously, `Bucket` would be the most minimal wrapper for a `Vec`, only
meant to hide those `Vec`s from the public interface. We did not mandate
that its items were sorted.

This initial decision meant that a `BucketAccumulator` would accept
unsorted buckets which needed to be sorted in the implementation via the
conversion to `SortedBucket`. In multithreaded accumulation, this meant
that the sorting would be performed while holding a lock on the inner
`BucketAccumulator`, which was less than ideal.

This change makes the critical section shorter (w.r.t. the time the lock
is held) by mandating that a `Bucket` already contains sorted items. The
conversion to a `SortedBucket` thus amounts to a simple re-packaging.