benches: add stub for benchmark comparing sortbuf to alternatives
`sortbuf` provides functionality already present in `std` in some way.
Our claim is (or will be) that our implementation provides some benefits
while not regressing (too much) on performance.

Before making any claims regarding performance, we should measure the
actual performance. Since we are intrested in how well we fare in
comparison to other approaches, we make our benchmark compare the
performance of sortbuf vs. those approaches.

The stud introduced by this change will be developed into such a
comparing benchmark. As a single measurement will take a considerable
amount of time, we can't use frameworks such as `criterion` based on a
statistcal analysis of multiple runs. On the other hand, this shouldn't
be necessary as the lenghy execution time of a single run should already
contribute to a low relative deviation. And we are, for now, only
interested in ballpark numbers anyway.