extender: replace explicit &mut BucketAccumulator impls with blanket
Previously, the `BucketAccumulator` trait would be implemented not for
`SortBuf`, and various lock types but mutable references on them. This
would safe some users from "loosing" an instance in a newly created
`Extender`. However, it also made the impls for lock types useless since
the lock would have to guard a `&mut SortBuf` (or some other
non-existant) `BucketAccumulator`.

This change makes these impls usable by implementing `BucketAccumulator`
for `SortBuf` itself and adding a blanket impl for mutable references.
Now users may "loose" a value in an `Extender`, but we deem the
compiler's reporting and solution obvious enough to be a non-issue.