benches: add fns for creating an Iterator over random items
In our comparison benchmark, we want to compare performance metrics of
different approaches for sorting large amounts of data. In order to do
so, we need to generate such data. While the generation should be fast
if should also yield sufficiently random data.

We already use a PCG implementation for our tests which is supposedly
fast in generating random values. We decided to use the same generation
approach for our benchmarks. A dummy generate-and-discard benchmark will
help assert this to some extend by providing a baseline.