SortBuf: store buckets in a Vec, heapify in IntoIterator::into_iter
Up until now, `SortBuf` wrapped a `BinaryHeap` of `SortedBucket`s. While
this meant a small additional cost when adding a bucket, the impl of
`IntoIterator::into_iter` amounted to a repackaging of that heap.

Performance-wise, whether we potentially propagate newly inserted
buckets through the heap or heapify an entire `Vec` should make little
difference. The latter one _might_ be a tad faster, at least in single-
threaded use-cases. More importantly, storing the buckets in a `Vec`
allows simpler access of individual buckets. This in turn, allows us to
re-distribute items in order to increase the drain speed and emptying
buckets early.

We may or may not end up introducing such a re-distribution step in the
future. However, enabling such a change now doesn't hurt. And as written
above, it will likely have no measurable effects on performance.