iter: drop From<BinaryHeap> impl for Iter in favor for From<Vec>
We recently made `SortBuf` wrap a `Vec` instead of a `BinaryHeap` and
switched to the new `From<Vec>` for creating an `Iter` in the process.
The only places where we explicitly create an `Iter` from `BinaryHeap`s
left are in `tests`, and those are easily migrated over, even saving
some lines.