bucket: make DEFAULT_BUCKET_BYTESIZE a standalone constant
We initially introduced `DEFAULT_BUCKET_BYTESIZE` as a constant to use
inside `Extender::with_default_bucket_size`. Back then we decided to
expose it to library users both for documentation purposes and as a
basis for alternative bucket sizes. However, we still made that constant
a member of `Extender`.

This turned out to be a suboptimal decision: as `Extender` is generic
over `BucketAccumulators`, using the constant wasn't as simple as
writing `Extender::DEFAULT_BUCKET_BYTESIZE` but required specifying the
type parameter.

This change makes the constant more accessible by moving it outside
`Extender`. Normally, it would thus remain inside the `extender` module.
As we did not find this an appropriate place for this constant, we
decided to move it over to `bucket`. This in turn meant we had to touch
the documentation of `DEFAULT_BUCKET_BYTESIZE` due to it previously
referencing the `BucketAccumulator` trait.