benches: add SPDX-header to comparison benchmark
When introducing the benchmark, we omitted any licensing information. We
half forgot about it, but we also didn't care too much. But on second
thought, the way we wrote the benchmark may be interesting for others
hitting similar problems for other algorithms or data structures. Adding
licensing info makes re-use less awkward.