bucket: add fn for determining a SortedBucket's overcapacity
Currently, `Vec::pop` apparently doesn't feature any logic for shrinking
the underlying storage. This is probably intentional in order to avoid
unwanted (run time) costs.

Thus, with our current code, we'll occupy all the memory allocated for a
bucket until it is drained and removed from `Iter`'s internal heap.
Assuming a sufficiently uniform distribution of items among buckets,
this will cause most of the memory to be freed only after an `Iter` is
almost completely drained. Considering that the data structure is meant
for large amounts of data, this is suboptimal.

We do want to free some memory earlier during iteration. We can do so by
shrinking the underlying `Vec`s. Shrinking after each single item
extraction would be possible but potentially inefficient. We want to
avoid calling all too often into the allocator. Instead, we wan to
perform a shrink only if we have some amount of unused memory.

This change introduces a function for determining the number of unused
slots for items in the `SortedBucket`.