bucket: add fn for shrinking a SortedBucket
Currently, `Vec::pop` apparently doesn't feature any logic for shrinking
the underlying storage. This is probably intentional in order to avoid
unwanted (run time) costs.

Thus, with our current code, we'll occupy all the memory allocated for a
bucket until it is drained and removed from `Iter`'s internal heap.
Assuming a sufficiently uniform distribution of items among buckets,
this will cause most of the memory to be freed only after an `Iter` is
almost completely drained. Considering that the data structure is meant
for large amounts of data, this is suboptimal.

We do want to free some memory earlier during iteration. We can do so by
shrinking the underlying `Vec`s. We could do so in the `Iterator::next`
impl for `SortedBucket`. However, if we want to make the amount of
unused memory a bucket is allowed to accumulate before shrinking
configurable via `Iter`. Thus, we only provide a function for shrinking
which will be called from `Iter`.