extender: don't push empty buckets when dropping an Extender
When an `Extender` is dropped, we need to move any residual data to the
`BucketAccumulator`. This step could involve taking a lock, which could
block, for some impls of `BucketAccumulator`.

If there is no data in the accumulator, we can skip this potentially
expensive step. Incidentially, by not pushing empty buckets to the
accumulator, we also make the iteration an unmeassuably tiny bit faster,
since those don't have to be popped from the heap.