extender: replace "with_*" constructors with new + "set_*" fns
When introducing the `Extender` type, we included three different
construction methods, allowing specification of the target bucket size
by items or bytes or choosing a default size. This approach seemed
reasonable at the time, but made construction awkward due to the long
function names. It also effectively obstructed the introduction of a
"simple" function to `BucketAccumulator` for creating an extender.

We thus decided to switch to a plain constructor named `new` alongside
two setters for the target bucket size. These do, of course, introduce
the possibility to alter the target bucket size for an existing
`Extender`, which we did try to avoid. However, we deemed the
possibility to not be hazardeous in any way. We'd still choose an
builder-llike interface consumind and returning an `Extender` rather
than a mutable reference, but those are obstructed by the existing
`Drop` impl.