extender: make BucketAccumulator::add_bucket communicate errors
Previously, we'd simply let existing implementations of that function
panic on (re-)allocation failures. However, with a data structure meant
for large amounts of data, allocation failures are more common than
you'd generally expect.

We want to offer users opportunities to cope with allocation failures.
We thus report an error back to the caller. Since we don't want to lose
any data, we return that error alongside the `Bucket` which could not be
returned. This allows users to re-try the operation when there is memory
availible again.