extender: add fn for inserting items
The `Extend` impl for `Extender` used to be the only interface through
which users could push items to the buffer. Unfortunately, `Extend` does
not provide any opportunity to communicate errors to callers, and we do
want to communicate (re-)allocation failures rather than panic.

This new function serves as an interface with such error reporting
capabilities, and will be the canonical interface to push items in the
future.

Besides the error reporting, the function also takes another approach
than the existing `Extend` impl. We do not use `BucketGen`, which we
originally only introduced in order to simplify committing multiple
buckets at once --an appraoch we abandoned because of its implications
on critical sections. We also try to fill buckets to their full capacity
rather than to the target bucket size. The item accumulator's capacity
may end up being higher than the bucket size depending on the allocator.