extender: migrate Extend impl for Extender to Extender::insert_items
We recently introduced `Extender::insert_items`, a function for
interting items featuring error reporting and making better use of
memory, as a future canonical interface. To reduce code duplication, we
do waway with the existing logic in the `Extend::extend` impl for
`Extender` and just use said `insert_items`.

This also removes the only user of the obsolete `BucketGen` type, which
will be removed by a future change.