extender: remove now obsolete BucketGen
Originally, we thought it would be a good idea to commit as many
`Bucket`s as possible in one go to a given `BucketAccumulator`. In
particular, we decided to hand over an `Iterator` which would be drained
in the body of the receiving fn `BucketAccumulator::add_buckets`.
`BucketGen` was introduced to fill that `Iterator`'s role.

However, that approach meant that all the actual work of accumulating
items into buckets, including sorting the items within a bucket, would
happen inside `BucketAccumulator::add_buckets`. Since we planned to
allow pushing items from multiple threads from the very start, that
meant that all the work would happen inside the critical section. The
idea was stupid.

We then changed the design and replaced `BucketAccumulator::add_buckets`
with `BucketAccumulator::add_bucket`, which would only add a single
`Bucket` that would already be accumulated and sorted. However, we'd
still use `BucketGen` inside `Extender` for constructing buckets from
items.

Recently, we tweaked both `BucketAccumulator` and `Extender` again in
order to allow for allocation errors to be reported. Rather than having
`BucketGen` yield `Result`s we decided to duplicate the logic in a new
function `Extender::insert_items` and migrate the existing `Extend` impl
to use that function. This left `BucketGen` unused.

This change now removes this type, which was motivated by a non-ideal
design decision which was already remedied a while ago.