bucket: add fn for converting a Bucket back to a Vec
We construct `Bucket`s from a `Vec` serving as an accumulator for items
and commit it to a `BucketAccumulator`. Recently, made the comitting
function return the bucket back to the caller in the case of some
failures. It we want to preserve the option of comitting it again, we
need to hold onto it.

The most straightforward way to do this would be to make it as if the
committing never happened, i.e. have the bucket turned back into a
`Vec` and make it the item accumulator. However, we previously lacked
the possibility to turn a `Bucket` back into a `Vec`. This change
introduces a function for doing so.

We could have implemented `From<Bucket>` for `Vec`, but we did not want
to expose this funcitonality to users of the library.