extender: rename Extender to Inserter
So we have this type named `Extender`, which got its name primarily from
the fact that it implements `Extend` and commits items to a buffer.
While working on this library, we got more and more dissatisfied with
the name. Naming things sure is hard.

While writing documentation and commit messages, we often found
ourselves writing about the "insertion" of items. And somewhat recently,
we added a fn `Extender::insert_items`. Again, "insert". `Inserter` does
appears to be the more natural name for the thing. Which is why we now
rename it.

This change does only rename the type, but does so in all instances
where it matters, including documentation and test names. Future changes
will also eliminate the name "extender" in other contexts, e.g. as
variable or module name.