bucket: impl Debug for Bucket
It's common in rust to implement `Debug` for public types to facilitate
debugging on the API consumer's side. The recommended way to implement
`Debug` is via `derive(Debug)`. For structs, that will result in an
implementation formatting all members.

`Bucket` is a datatype which is meant to hold significant amounts of
data. Printing that data in any form won't benefit developers in any
way but just swamp their terminals with megabytes worth or data. Hence,
`derive(Debug)` is not an option. Instead, we provide a handwritten
impl.