bucket: impl Debug for SortedBucket
It's common in rust to implement `Debug` for public types to facilitate
debugging on the API consumer's side. The recommended way to implement
`Debug` is via `derive(Debug)`. For structs, that will result in an
implementation formatting all members.

`SortedBucket` is not a public data type, but implementing `Debug` for
it allows us to apply `derive(Debug)` to `SortBuf` and `Iter`. However,
like a `Bucket`, a `SortedBucket` will likely hold significant amounts
of data. Thus, the output of a `Debug` impl via `derive(Debug)` provides
little utility. We therefore provide a handwritten impl.