inserter: let Inserter::insert_items return an Error without an Iterator
When introducing support for failable allocation, this included
`Inserter::insert_items`. In order to allow for reaction to allocation
failures without any data-loss, we'd return an `Iterator` over all items
that could not be inserted alongside the allocaiton failure.

That decision, however, made calling the function more awkward than
necessary as it forced handling that `Iterator` in some way. Since that
`Iterator` would not implement `Debug`, or rather, we could not promise
that it would, we even had to filter out that `Iterator` when "handling"
the error by means of `Result::expect` or `Result::unwrap`.

The utility of the `Iterator` returned in the error case was lacking as
well: as its type would differ from the `Iterator` fed to the funciton,
we'd need a different type for each re-try or a boxed iterator, which
would be both slow and problematic in the face of a previous allocation
failure.

At the same time, a more practical solution for resumption already
existed with the possibility to feed a `&mut Iterator` rather than an
`Iterator` by value (e.g. the return value of `Iterator::by_ref`). Since
we returned the partially drained `Iterator`, the existing
implementation of `Inserter::insert_items` already was written so that
only items were pulled from that `Iterator` for which we already had
memory allocated.

So in the end, the extra value returned on error contributed littel to
nothing but made calls to the function more ugly. We therefore decided
to do away with it.